wip: transfer scheduling creation

+++ src/main/java/com/Isabela01vSilva/bank_isabela/service/client/dto/PayloadDTO.java
@@ -2,7 +2,7 @@ package com.Isabela01vSilva.bank_isabela.service.client.dto;
 
 import java.math.BigDecimal;
 
-public record PayloadDTO(BigDecimal valor,
+public record PayloadDTO(Double valor,
                          String numeroContaOrigem,
                          String numeroContaDestino) {
 }
